r/storage_table: moving the table scripts into the migration folder
